# backend/internal/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_studentinovisad_popisomator_backend_internal_dto.AddUpdateItemTypePropertyRequest:
    properties:
      default_value:
        type: string
      property_id:
        type: integer
      type_id:
        type: integer
      visibility:
        allOf:
        - $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_repository.PropertyVisibility'
        enum:
        - overview
        - details
    required:
    - property_id
    - type_id
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_dto.CreateItemTypeRequest:
    properties:
      derived_name_format:
        type: string
      description:
        type: string
      name:
        type: string
      properties:
        items:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemTypeProperty'
        type: array
    required:
    - name
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_dto.ItemType:
    properties:
      derived_name_format:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      properties:
        items:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemTypeProperty'
        type: array
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.ItemTypeOption:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.ItemTypeProperty:
    properties:
      default_value:
        type: string
      id:
        type: integer
      visibility:
        enum:
        - overview
        - details
        type: string
    required:
    - id
    type: object
  github_com_studentinovisad_popisomator_backend_internal_dto.ItemTypesPage:
    properties:
      items:
        description: ItemType page items
        items:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemType'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_dto.UpdateItemTypeRequest:
    properties:
      derived_name_format:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - id
    type: object
...
  github_com_studentinovisad_popisomator_backend_internal_repository.PropertyVisibility:
    enum:
    - overview
    - details
    type: string
    x-enum-varnames:
    - PropertyVisibilityOverview
    - PropertyVisibilityDetails
...
  github_com_studentinovisad_popisomator_backend_internal_response.Error:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: Inventory management API for Popisomator.
  title: Popisomator API
  version: "1.0"
paths:
...
  /item-types:
    get:
      parameters:
      - description: Page size (default 20, max 50)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemTypesPage'
        "400":
          description: invalid limit/offset
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: List item types
      tags:
      - ItemTypes
    post:
      consumes:
      - application/json
      parameters:
      - description: Item type to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.CreateItemTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemType'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Create an item type (manager/admin only)
      tags:
      - ItemTypes
  /item-types/{id}:
    delete:
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: invalid type id
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Delete an item type (manager/admin only)
      tags:
      - ItemTypes
    get:
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemType'
        "400":
          description: invalid type id
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Get an item type by ID
      tags:
      - ItemTypes
    patch:
      consumes:
      - application/json
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.UpdateItemTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemType'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Update an item type (manager/admin only)
      tags:
      - ItemTypes
  /item-types/{id}/properties:
    post:
      consumes:
      - application/json
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Property to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.AddUpdateItemTypePropertyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemTypeProperty'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Add a property to an item type (manager/admin only)
      tags:
      - ItemTypes
  /item-types/{id}/properties/{prop_id}:
    delete:
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Property ID
        in: path
        name: prop_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: invalid type/property id
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Remove a property from an item type (manager/admin only)
      tags:
      - ItemTypes
    patch:
      consumes:
      - application/json
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Property ID
        in: path
        name: prop_id
        required: true
        type: integer
      - description: Property fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.AddUpdateItemTypePropertyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemTypeProperty'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: Update a property on an item type (manager/admin only)
      tags:
      - ItemTypes
  /item-types/options:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_dto.ItemTypeOption'
            type: array
        "401":
          description: not logged in
          schema:
            $ref: '#/definitions/github_com_studentinovisad_popisomator_backend_internal_response.Error'
      security:
      - CookieAuth: []
      summary: List all item types in a minified form, for dropdowns
      tags:
      - ItemTypes
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"